"debug client refactored into more layers; updates #832"
